Changed the calculation of KLD scores, calculate_kld.
Introduced and error in which kld scores were handled as strings rather than floats, so sorting and selection did not work as expected. That fix is now squashed here. Leaving message in commit just in case since debugging this took a while the first time.
Reflected changes to function in execute script.