Changes to function names which read/write data.
Added a small section header for user defined functions.
Note that the change from sys.argv to variables is reflected in a later commit. For some reason, could not reorder or squash that particular commit. Variable now finds data files in its path, but can be easily reverted to a sys.argv if desired.