Changed assignment and saving of array ids into a separate function.
Modified the functions where it was being used (i.e. I removed the lines of code which are identical to the function and passed in the dictionary form of the array ids instead). Note that in these functions I also removed the arbitrary conditions regading whether or not the array name was less than 40 characters. Instead, the array id is not used every time. Necessary changes to execute script also included.