Removed all instances of plotting.
Removed plotting of adsorption data and subseuqent fit from analyze_all_ratios function.
Removed the summary plot of all Henry's coefficients from the end of 'execute_henrys_coefficient_analysis.py'.

The latter call was actually broken to begin with, as the file containing the plotting function was never imported. This commit is atomic by itself, and makes the previous commit atomic, but per the reason in the previous commit message, was done as a separate commit.