Removed plotting functions and moved subset to a separate file.
N.B. This commit is NOT atomic, as there are still instances in which these functions are called in the scripts 'henrys_coefficient_analysis.py' and 'execute_henrys_coefficient_analysis.py', however these are all removed in the susequent commit.

The reason for doing this as two separate commits is that the one call in 'execute_henrys_coefficient_analysis.py' was actually broken to begin with, as the function was never imported. The breaking of this file occured as part of the previous PR, but was not caught.